requirements.txt
astroid==3.3.5
coverage==7.6.4
dill==0.3.9
exceptiongroup==1.2.2
Faker==30.8.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tabulate==0.9.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
